requirements.txt
beautifulsoup4
python-dateutil
python-dotenv
faiss-cpu
numpy
tzlocal
google-auth
google-auth-oauthlib
google-api-python-client
langchain-huggingface
mistralai
sentence-transformers
qdrant-client
PyPDF2
openpyxl
pandas
requests
streamlit
